Moved IconSelector to own component
